+++ ChouMarin/Sources/Sensor/SensorBase.hpp
@@ -69,7 +69,7 @@ protected:
 
 public:
   SensorBase();
-  SensorBase(SensorType sT, DataType dT, unsigned int id);
+  SensorBase(SensorType sT, DataType dT, short unsigned int id);
 
   virtual ~SensorBase(){};
   virtual const SensorData &getData() = 0; // default value set to 0
